refactor: move qid_to_label import from cal_util to calibration.util
