+++ src/qdash/workflow/tasks/benchmark/x180_interleaved_randoized_benchmarking.py
@@ -1,7 +1,7 @@
 from typing import ClassVar
 
 from qdash.datamodel.task import DataModel
-from qdash.workflow.cal_util import qid_to_label
+from qdash.workflow.calibration.util import qid_to_label
 from qdash.workflow.tasks.base import (
     BaseTask,
     InputParameter,
