+++ src/qdash/workflow/tasks/measurement/readout_classification.py
@@ -3,7 +3,7 @@ from typing import TYPE_CHECKING, ClassVar
 if TYPE_CHECKING:
     import plotly.graph_objs as go
 from qdash.datamodel.task import DataModel
-from qdash.workflow.cal_util import qid_to_label
+from qdash.workflow.calibration.util import qid_to_label
 from qdash.workflow.tasks.base import (
     BaseTask,
     OutputParameter,
