+++ src/qdash/workflow/tasks/one_qubit_coarse/check_t1.py
@@ -2,7 +2,7 @@ from typing import ClassVar
 
 import numpy as np
 from qdash.datamodel.task import DataModel
-from qdash.workflow.cal_util import qid_to_label
+from qdash.workflow.calibration.util import qid_to_label
 from qdash.workflow.tasks.base import (
     BaseTask,
     InputParameter,
