+++ src/qdash/workflow/tasks/one_qubit_coarse/chevron_pattern.py
@@ -1,7 +1,7 @@
 from typing import ClassVar
 
 import numpy as np
-from qdash.workflow.cal_util import qid_to_label
+from qdash.workflow.calibration.util import qid_to_label
 from qdash.workflow.tasks.base import (
     BaseTask,
     InputParameter,
